pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["DayaneBatista <nbdayane12@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12"
flask = "^3.0.0"
gunicorn = "^21.2.0"
flask-sqlalchemy = "^3.1.1"
werkzeug = "^3.1.3"
pytesseract = "^0.3.10"
pillow = "^10.3.0"
opencv-python = "^4.9.0.80"
pdf2image = "^1.16.3"
PyPDF2 = "^3.0.1"
spacy = "^3.7.4"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.23.0"
numpy = "^2.3.1"
pandas = "^2.1.4"


[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"